Add new recipe to setup the full Sysinternals Suite
Add recipe sysinternals::full_suite to setup the complete sysinternals
suite at once.
This recipe reuse the `sysinternals.install_dir` attribute and add
two new attributes to control the suite url and its checksum:
* `sysinternals.full_suite.url`
* `sysinternals.full_suite.checksum`

By default checksum is nil to skip the checksum validation, but it
is a good practice to perform this validation.

Signed-off-by: Baptiste Courtois <b.courtois@criteo.com>